Move authentication parser into the http/parse module
