+++ src/http/parse.rs
@@ -5,12 +5,20 @@
 //! not a value, use [ContentType]. If you want to parse dynamically based on
 //! the response's metadata, use [ContentType::parse_response].
 
-use crate::http::Response;
+use crate::{collection::Authentication, http::Response, template::Template};
 use anyhow::{anyhow, Context};
 use derive_more::{Deref, Display, From};
 use regex::Regex;
 use serde::{de::IntoDeserializer, Deserialize, Serialize};
 use std::{borrow::Cow, ffi::OsStr, fmt::Debug, path::Path, sync::OnceLock};
+use base64::{prelude::BASE64_STANDARD, Engine};
+use std::str;
+use nom::{
+    bytes::complete::{tag, take_until},
+    sequence::pair,
+    IResult,
+};
+
 
 /// All supported content types. Each variant should have a corresponding
 /// implementation of [ResponseContent].
@@ -163,6 +171,54 @@ impl ContentType {
     }
 }
 
+impl Authentication {
+    /// Convert the value of an Authorization header into an authentication struct
+    /// Can either be Bearer or Basic
+    pub fn from_header_value(input: &str) -> anyhow::Result<Self> {
+        fn bearer(input: &str) -> IResult<&str, &str> {
+            tag("Bearer ")(input)
+        }
+
+        fn basic(input: &str) -> IResult<&str, &str> {
+            tag("Basic ")(input)
+        }
+
+        fn username_and_password(input: &str) -> IResult<&str, &str> {
+            let (password, (username, _)) =
+                pair(take_until(":"), tag(":"))(input)?;
+            Ok((username, password))
+        }
+
+        if let Ok((token, _)) = bearer(input) {
+            let parsed: Template =
+                token.to_string().try_into().map_err(|_| {
+                    anyhow!("Cannot convert auth header to template")
+                })?;
+            return Ok(Self::Bearer(parsed));
+        }
+
+        if let Ok((encoded, _)) = basic(input) {
+            let decoded_bytes = BASE64_STANDARD.decode(encoded)?;
+            let decoded = str::from_utf8(decoded_bytes.as_slice())?;
+
+            let (username, password) =
+                match username_and_password(decoded) {
+                    Ok((u, p)) => (
+                        u.to_string().try_into()?,
+                        Some(p.to_string().try_into()?),
+                    ),
+                    Err(_) => (decoded.to_string().try_into()?, None),
+                };
+
+            return Ok(Self::Basic { username, password });
+        }
+
+        return Err(anyhow!("Failed to parse auth header"));
+    }
+}
+
+
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -286,5 +342,37 @@ mod tests {
         let mut headers = HeaderMap::new();
         headers.insert(CONTENT_TYPE, content_type.try_into().unwrap());
         headers
+    }
+
+    #[test]
+    fn parse_auth_header_test() {
+        let example = "Basic Zm9vOmJhcg==";
+        match Authentication::from_header_value(example).unwrap() {
+            Authentication::Basic { username, password } => {
+                assert_eq!(username.to_string(), "foo");
+                assert_eq!(password.unwrap().to_string(), "bar");
+            }
+            _ => panic!("Should be basic auth!"),
+        };
+
+        let example = "Basic dXNlcm5hbWV3aXRob3V0cGFzc3dvcmQ=";
+        match Authentication::from_header_value(example).unwrap() {
+            Authentication::Basic { username, password } => {
+                assert_eq!(username.to_string(), "usernamewithoutpassword");
+                assert!(password.is_none());
+            }
+            _ => panic!("Should be basic auth!"),
+        };
+
+        let example = "Bearer eyjlavljhhkjasdjlkhskljdfklasdlkjhf";
+        match Authentication::from_header_value(example).unwrap() {
+            Authentication::Bearer(bearer) => {
+                assert_eq!(
+                    bearer.to_string(),
+                    "eyjlavljhhkjasdjlkhskljdfklasdlkjhf"
+                )
+            }
+            _ => panic!("Should be bearer auth!"),
+        }
     }
 }
